Associate mod list files with the executable instead of a DLL
Reorganizing the program to be lots of small files instead of one monolithic
executable changed the output of System.Reflection.Assembly.GetExecutingAssembly
to Wabbajack.dll instead of Wabbajack.exe.

This fix assumes the DLL containing the calling code is in the same folder
as Wabbajack.exe.  There's room for improvement if anyone can figure out how
to positively identify Wabbajack.exe.