Use a different Google Drive link to verify links during compile
Using the export link to compile is kind of unreliable as the
arbitrary traffic control Google implements will affect how things
work.  Using the /file/d/ link still gives a pretty solid indication
that the file is valid and available while not needing to be able
to download the file at compile time.