Fix randomness of lists presented
The list of mod lists was being sorted two different times.
The second time shuffled the list and then sorted using a method
that made the list very consistently show the same order of mod
lists.  I think it was in order of most recently updated...?

Now only one sort is performed and the list is properly shuffled
each time the program is launched.