Fix compiling with OMODs
Compiling with Oblivion OMODs was failing with the error "Unable to
cast object of type 'Wabbajack.Common.RelativePath' to type
'Wabbajack.Common.AbsolutePath'."

Previously, GatheringExtractWithOMOD was creating ExtractedNativeFiles
by providing the absolute path of the file and the name of the subfolder
(e.g., "C:\downloads\0123\0123_tmp" and "0123_tmp").  This caused the error
seen as a relative path of "0123_tmp" couldn't be converted to an
absolute path.

Now, only the absolute path is provided which seems to work fine.