Add error reporting to the launcher
This reports errors to the user and exits the launcher but
ONLY if a previous version of WJ could not be found. Existing
users will fail silently and be unable to update. New users
will get some kind of error they can report to the team without
the launcher getting stuck.